# core-lending-provisioning-risk-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-provisioning-risk
  description: Provisioning Risk Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-origination
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/provisioning-cases/{provisioningCaseId}/stage-history/{provisioningStageHistoryId}:
    get:
      tags:
        - ProvisioningStageHistory
      summary: Get a stage history entry by ID
      operationId: getById_1
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningStageHistoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningStageHistoryDTO'
    put:
      tags:
        - ProvisioningStageHistory
      summary: Update a stage history entry
      operationId: update_1
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningStageHistoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisioningStageHistoryDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningStageHistoryDTO'
    delete:
      tags:
        - ProvisioningStageHistory
      summary: Delete a stage history entry
      operationId: delete_1
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningStageHistoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/provisioning-cases/{provisioningCaseId}/stage-history:
    get:
      tags:
        - ProvisioningStageHistory
      summary: List or search stage history for a provisioning case
      operationId: findAll_1
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.provisioningCaseId
          in: query
          description: Exact filter for provisioning case id
          required: false
          schema:
            type: string
        - name: filters.oldStageCode
          in: query
          description: Exact filter for old stage code
          required: false
          schema:
            type: string
            enum:
              - STAGE_1
              - STAGE_2
              - STAGE_3
              - POCI
        - name: filters.newStageCode
          in: query
          description: Exact filter for new stage code
          required: false
          schema:
            type: string
            enum:
              - STAGE_1
              - STAGE_2
              - STAGE_3
              - POCI
        - name: filters.eclAmountAtChange
          in: query
          description: Exact filter for ecl amount at change
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[eclAmountAtChange].from
          in: query
          description: Filter ecl amount at change from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[eclAmountAtChange].to
          in: query
          description: Filter ecl amount at change to value
          required: false
          schema:
            type: string
        - name: filters.changedAt
          in: query
          description: Exact filter for changed at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[changedAt].from
          in: query
          description: Filter changed at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[changedAt].to
          in: query
          description: Filter changed at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.changedBy
          in: query
          description: Exact filter for changed by
          required: false
          schema:
            type: string
        - name: filters.reason
          in: query
          description: Exact filter for reason
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseProvisioningStageHistoryDTO'
    post:
      tags:
        - ProvisioningStageHistory
      summary: Create a new stage history entry
      operationId: create_1
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisioningStageHistoryDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningStageHistoryDTO'
...
components:
  schemas:
...
    ProvisioningStageHistoryDTO:
      required:
        - changedAt
        - changedBy
        - eclAmountAtChange
        - newStageCode
        - oldStageCode
        - provisioningCaseId
        - reason
      type: object
      properties:
        provisioningStageHistoryId:
          type: string
          format: uuid
          readOnly: true
        provisioningCaseId:
          type: string
          format: uuid
        oldStageCode:
          type: string
          enum:
            - STAGE_1
            - STAGE_2
            - STAGE_3
            - POCI
        newStageCode:
          type: string
          enum:
            - STAGE_1
            - STAGE_2
            - STAGE_3
            - POCI
        eclAmountAtChange:
          type: number
        changedAt:
          type: string
          format: date-time
        changedBy:
          maxLength: 100
          minLength: 0
          type: string
        reason:
          maxLength: 500
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseProvisioningStageHistoryDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ProvisioningStageHistoryDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...